# statistics: Improve performance for histogram data creation